// admin-clients/libs/shared/venues/feature/standard-venue-tpl-editor/src/lib/models/venue-tpl-editor-modes.enum.ts
export enum InteractionMode {
    all = 'all',
    interactive = 'int',
    graphic = 'gra'
}

export enum EditorMode {
    base = 'base',
    seatMatrixCreate = 'seatMatrixCreate',
    seatMatrixIncrease = 'seatMatrixIncrease',
    blocksSetup = 'blocksSetup',
    weightsSetup = 'weightsSetup'
}

export enum VisualizationMode {
    blocks = 'blocks',
    weights = 'weights',
    names = 'names'
}